color_palette/requirements.txt
Pillow
plotly
beautifulsoup4
